Fix the "unknown" egg versions problem in deploy-reqs
Reverts PR #21 and fixes issue #44.

Instead of using an alternate src dir, now we just ignore the folders
that don't look like egg dirs. The only know case for this is when we
try to install packages from git and the src dir is left in the same
folder as the eggs folders. So, it shouldn't be an issue.