Ignore errors on tmp dir removal on shub deploy
A race condition was reported on windows platforms, see issue #58 and
the error on the tmp dir removal was masquerading the real error.

As these folders are usually pretty small and the error only rarely happens on
windows I think that just ignoring the error won't cause trouble to users.