Make `get_docker_client` compatible with old and new versions of docker-py
The docker Client class is now called DockerClient in newer versions of the docker python
package.

There's also the problem that the `docker-py` no longer is called docker-py, only `docker`.
So now if you install have the docker-py installed, it will obfuscurate the new version of the package,
and you are advised to remove it and make `pip install docker` again.

That's why I also changed the package name in the requirements